+++ windows/src/TinyClips.Core/Services/ServiceCollectionExtensions.cs
@@ -12,7 +12,6 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IFileSystem, FileSystem>();
         services.AddSingleton<IFileNameService, FileNameService>();
         services.AddSingleton<IClipStorageService, ClipStorageService>();
-        services.AddSingleton<IClipsLibraryService, ClipsLibraryService>();
         services.AddSingleton<IHotKeyService, HotKeyService>();
         services.AddSingleton<ILaunchAtLoginService, LaunchAtLoginService>();
         services.AddSingleton<IAudioDeviceService, AudioDeviceService>();